LayoutTests/fast/page-color-sampling/color-sampling-ignores-non-fixed-content-expected.txt

PASS colorBeforeFading.top is "rgb(255, 100, 0)"
PASS colorAfterFading.top is "rgb(255, 100, 0)"
PASS successfullyParsed is true

TEST COMPLETE
